修复自动生成模型时 DECIMAL 类型，设置小数不起作用的问题

+++ app/service/sys/generate.js
@@ -76,7 +76,7 @@ class GenerateService extends Service {
         if (v.autoIncrement) {
           autoIncrement = `, autoIncrement:${v.autoIncrement}`;
         }
-      } else if (v.type === 'FLOAT' || v.type === 'DOUBLE' || v.type === 'DOUBLE') {
+      } else if (v.type === 'FLOAT' || v.type === 'DOUBLE' || v.type === 'DECIMAL') {
         const tt = `type: DataTypes.${v.type}`;
         if (v.unsigned && v.zerofill) {
           type = `${tt}.UNSIGNED.ZEROFILL`;
